Add files via upload
This lesson took me more than 40 minutes to finish because of the formatting of dates. I had a list that contained two dates I wanted to filter by in the format: 'm/dd/yy'. Nothing was matching the records in the CSV. In the end, it was due to the formatting of the year. It was in 4 digits, no the 2 year I had in my list.